+++ quantify_scheduler/operations/pulse_factories.py
@@ -6,7 +6,16 @@ A module containing factory functions for pulses on the quantum-device layer.
 These factories are used to take a parametrized representation of on a operation
 and use that to create an instance of the operation itself.
 """
+import numpy as np
+
+from quantify_scheduler.backends.qblox import constants as qblox_constants
+from quantify_scheduler.backends.qblox import helpers as qblox_helpers
 from quantify_scheduler.operations import pulse_library
+from quantify_scheduler.operations.stitched_pulse import (
+    StitchedPulse,
+    StitchedPulseBuilder,
+)
+from quantify_scheduler.resources import BasebandClockResource
 
 
 def rxy_drag_pulse(
@@ -134,3 +143,194 @@ def nv_spec_pulse_mw(
         clock=clock,
         port=port,
     )
+
+
+def long_square_pulse(
+    amp: float,
+    duration: float,
+    port: str,
+    clock: str = BasebandClockResource.IDENTITY,
+    t0: float = 0,
+) -> StitchedPulse:
+    """Create a long square pulse using DC voltage offsets.
+
+    .. note::
+
+        This function creates a
+        :class:`~quantify_scheduler.operations.stitched_pulse.StitchedPulse` object,
+        which can currently only be compiled by the Qblox backend.
+
+    Parameters
+    ----------
+    amp : float
+        Amplitude of the envelope.
+    duration : float
+        The pulse duration in seconds.
+    port : str
+        Port of the pulse, must be capable of playing a complex waveform.
+    clock : str, optional
+        Clock used to modulate the pulse. By default the baseband clock.
+    t0 : float, optional
+        Time in seconds when to start the pulses relative to the start time
+        of the Operation in the Schedule. By default 0.
+
+    Returns
+    -------
+    StitchedPulse
+        A StitchedPulse object containing an offset instruction with the specified
+        amplitude.
+    """
+
+    pulse = (
+        StitchedPulseBuilder(port=port, clock=clock, t0=t0)
+        .add_voltage_offset(path_0=amp, path_1=0.0, duration=duration)
+        .build()
+    )
+    return pulse
+
+
+def staircase_pulse(
+    start_amp: float,
+    final_amp: float,
+    num_steps: int,
+    duration: float,
+    port: str,
+    clock: str = BasebandClockResource.IDENTITY,
+    t0: float = 0,
+    grid_time_ns: int = qblox_constants.GRID_TIME,
+) -> StitchedPulse:
+    """Create a staircase-shaped pulse using DC voltage offsets.
+
+    This function generates a real valued staircase pulse, which reaches its final
+    amplitude in discrete steps. In between it will maintain a plateau.
+
+    .. note::
+
+        This function returns a
+        :class:`~quantify_scheduler.operations.stitched_pulse.StitchedPulse` object,
+        which can currently only be compiled by the Qblox backend.
+
+    Parameters
+    ----------
+    start_amp : float
+        Starting amplitude of the staircase envelope function.
+    final_amp : float
+        Final amplitude of the staircase envelope function.
+    num_steps : int
+        The number of plateaus.
+    duration : float
+        Duration of the pulse in seconds.
+    port : str
+        Port of the pulse.
+    clock : str, optional
+        Clock used to modulate the pulse. By default the baseband clock.
+    t0 : float, optional
+        Time in seconds when to start the pulses relative to the start time
+        of the Operation in the Schedule. By default 0.
+    grid_time_ns : int, optional
+        Grid time in ns. The duration of each step of the staircase must be a multiple
+        of this. By default equal to the grid time of Qblox modules.
+
+    Returns
+    -------
+    StitchedPulse
+        A StitchedPulse object containing incrementing or decrementing offset
+        instructions.
+
+    Raises
+    ------
+    ValueError
+        When the duration of a step is not a multiple of ``grid_time_ns``.
+    """
+
+    builder = StitchedPulseBuilder(port=port, clock=clock, t0=t0)
+
+    try:
+        step_duration = (
+            qblox_helpers.to_grid_time(duration / num_steps, grid_time_ns) * 1e-9
+        )
+    except ValueError as err:
+        raise ValueError(
+            f"The duration of each step of the staircase must be a multiple of"
+            f" {grid_time_ns} ns."
+        ) from err
+    amps = np.linspace(start_amp, final_amp, num_steps)
+    for amp in amps:
+        builder.add_voltage_offset(path_0=amp, path_1=0.0, duration=step_duration)
+    pulse = builder.build()
+    return pulse
+
+
+def long_ramp_pulse(
+    amp: float,
+    duration: float,
+    port: str,
+    offset: float = 0,
+    clock: str = BasebandClockResource.IDENTITY,
+    t0: float = 0,
+    part_duration_ns: int = qblox_constants.STITCHED_PULSE_PART_DURATION_NS,
+) -> StitchedPulse:
+    """Creates a long ramp pulse by stitching together shorter ramps.
+
+    This function creates a long ramp pulse by stitching together ramp pulses of the
+    specified duration ``part_duration_ns``, with DC voltage offset instructions placed
+    in between.
+
+    .. note::
+
+        This function returns a
+        :class:`~quantify_scheduler.operations.stitched_pulse.StitchedPulse` object,
+        which can currently only be compiled by the Qblox backend.
+
+    Parameters
+    ----------
+    amp : float
+        Amplitude of the ramp envelope function.
+    duration : float
+        The pulse duration in seconds.
+    port : str
+        Port of the pulse.
+    offset : float, optional
+        Starting point of the ramp pulse. By default 0.
+    clock : str, optional
+        Clock used to modulate the pulse, by default a BasebandClock is used.
+    t0 : float, optional
+        Time in seconds when to start the pulses relative to the start time of the
+        Operation in the Schedule. By default 0.
+    part_duration_ns : int, optional
+        Duration of each partial ramp in nanoseconds, by default
+        :class:`~quantify_scheduler.backends.qblox.constants.STITCHED_PULSE_PART_DURATION_NS`.
+
+    Returns
+    -------
+    StitchedPulse
+        A ``StitchedPulse`` composed of shorter ramp pulses with varying DC offsets,
+        forming one long ramp pulse.
+    """
+    dur_ns = qblox_helpers.to_grid_time(duration)
+    num_whole_parts = (dur_ns - 1) // part_duration_ns
+    amp_part = part_duration_ns / dur_ns * amp
+    dur_left = (dur_ns - num_whole_parts * part_duration_ns) * 1e-9
+    amp_left = amp - num_whole_parts * amp_part
+
+    builder = StitchedPulseBuilder(port=port, clock=clock, t0=t0)
+
+    cur_offset = offset
+    for _ in range(num_whole_parts):
+        if not (np.isclose(offset, 0) and np.isclose(cur_offset, offset)):
+            builder.add_voltage_offset(path_0=cur_offset, path_1=0.0)
+        builder.add_pulse(
+            pulse_library.RampPulse(
+                amp=amp_part, duration=part_duration_ns * 1e-9, port=port
+            )
+        )
+        cur_offset += amp_part
+    if cur_offset != offset:
+        builder.add_voltage_offset(path_0=cur_offset, path_1=0.0)
+    builder.add_pulse(
+        pulse_library.RampPulse(amp=amp_left, duration=dur_left, port=port)
+    )
+
+    pulse = builder.build()
+
+    return pulse