+++ tests/scheduler/backends/qblox/operation_handling/empty_qasm_program.py
@@ -1,0 +1,12 @@
+"""Empty QASM program fixture."""
+import pytest
+
+from quantify_scheduler.backends.qblox.instrument_compilers import QcmModule
+from quantify_scheduler.backends.qblox.qasm_program import QASMProgram
+from quantify_scheduler.backends.qblox.register_manager import RegisterManager
+
+
+@pytest.fixture(name="empty_qasm_program_qcm")
+def fixture_empty_qasm_program():
+    """Empty QASMProgram object."""
+    yield QASMProgram(QcmModule.static_hw_properties, RegisterManager())